Moved app bar components to their own directory
